myapp/requirements.txt
Flask==3.1.1
google-cloud-firestore==2.21.0
